==> model-building/1_Data_modeling_cleaning.ipynb <==
# Autoload module updates
%load_ext autoreload
%autoreload 2

# Import modules
import numpy as np
import pandas as pd
import sklearn
import sys
sys.path.insert(0, '../')
import src.cleanup as cleanup
properties = pd.read_csv('..\data\cleaned_output.csv', skip_blank_lines=True)


# Read dataframe

# Data Cleaning 2.0
After analysing my data, I want to use it for predicts. Before building my model I need to further clean my data of
* Any NaN values
* Any string values
* Any duplicates

I have already confirmed that there are no duplicates, so the first thing to deal with will be strings, then NaNs.

Columns `id` and `type of sale`  need dropping since these contain useless data. The `state of the building` column will be converted to binary with fully renovated properties processed as 1s, and properties that need any sort of renovation processed as 0s.

columns = ['id','type_of-sale']
properties = cleanup.drop_column(properties, columns)

Now clean the renovation data:

# Replace specific value in a given column with new_value
def replace_col_value(df, column, value, new_value):
    for index, row in df.iterrows():
        if row[column] == value:
            df.at[index, column] = new_value
    return df

properties = cleanup.replace_col_value(properties, 'state_of_the_building', 'needs renovating', 0)
properties = cleanup.replace_col_value(properties, 'state_of_the_building', '0', 0)

renovated = ['just_renovated', 'as_new', 'good']
for i in renovated:
    properties = cleanup.replace_col_value(properties, 'state_of_the_building', i, 1)

cleanup.unique_df_values(properties)




Checking for NaNs:


properties.head()

Replace NaNs in terrace/garden with 1s if terrace area exists:

properties = cleanup.check_for_adjascent_data(properties, 'terrace_area', 'terrace', 1)
properties = cleanup.check_for_adjascent_data(properties, 'garden_area', 'garden', 1)
0

Replace other Garden and Terrace Nans with zeroes:

properties.head()
properties = cleanup.sort_by_column(properties, 'living_area')
properties.head()

Replace `furnished`, `open fire`, `terrace`, `terrace area`, `garden`, `garden area` and `swimming pool` nans with 0s.

columns = ['furnished', 'open_fire', 'terrace', 'terrace_area', 'garden', 'garden_area', 'swimming_pool']

for i in columns:
    properties = cleanup.replace_nan(properties, i)
properties.head()

Delete rows that are missing Kitchen Data

display(properties.shape)
properties = properties.dropna(subset=['fully_equipped_kitchen', 'total_land_area', 'total_property_area', 'number_of_facades'])

nan_counts = properties.isnull().sum()
print(nan_counts)

cleanup.column_type_float(properties)

properties.to_csv('..\data\cleaned_output_modeling.csv', index=False)
