==> model-building/2_Data_modeling.ipynb <==
# Autoload module updates
%load_ext autoreload
%autoreload 2

# Import modules
import numpy as np
import pandas as pd
import sklearn
import sys
sys.path.insert(0, '../')
import src.cleanupoutliers as cleanoutliers
properties = pd.read_csv('..\data\cleaned_output_2.csv', skip_blank_lines=True)
properties.head()

# Convert properties to a numpy dataframe
properties = cleanup.drop_column(properties, ['type of property', 'subtype of property', 'province', 'locality', 'postalCode'])
X = properties.drop(columns=['price']).to_numpy()
y = properties['price'].to_numpy()
display(y)

# Import train_test_split
# Apply train_test_split to dataframe

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=133)

from sklearn.linear_model import LinearRegression
regressor = LinearRegression()
regressor.fit(X_train, y_train)
print("The score of the training set is", regressor.score(X_train, y_train))
print("The score of the testing set is", regressor.score(X_test, y_test))

Below, I clean the data set of outliers.

# Manually clean outliers
# Set removal count to zero. Each clean will sum the amount of removed rows into removal_count


properties = cleanoutliers.clean_outliers(properties)

# Convert properties to a numpy dataframe
X = properties.drop(columns=['price']).to_numpy()
y = properties['price'].to_numpy()
display(y)

# Import train_test_split
# Apply train_test_split to dataframe

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=133)

# Fit and score model

from sklearn.linear_model import LinearRegression
regressor = LinearRegression()
regressor.fit(X_train, y_train)
print("The score of the training set is", regressor.score(X_train, y_train))
print("The score of the testing set is", regressor.score(X_test, y_test))

# Predictor
y_pred = regressor.predict(X_test)

# Check regressor coefficients
coefficients= regressor.coef_
print(coefficients)
# Check regressor intercept
intercept = regressor.intercept_
print(intercept)

# Manually create first prediction
print((X_test[0])*coefficients.sum()+ intercept)

# Check model's first prediction to compare vs. previous output
print(y_pred[0])
properties.head()

# Drop 'living area' from dataframe
properties = cleanup.drop_column(properties, 'living area')

# Redo and score the model
X = properties.drop(columns=['price']).to_numpy()
y = properties['price'].to_numpy()
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=133)
regressor.fit(X_train, y_train)
print("The score of the training set is", regressor.score(X_train, y_train))
print("The score of the testing set is", regressor.score(X_test, y_test))
