==> model-building/3_Data_modeling_houses.ipynb <==
# Autoload module updates
%load_ext autoreload
%autoreload 2

# Import modules
import numpy as np
import pandas as pd
import sklearn
import sys
sys.path.insert(0, '../')
import src.cleanup as cleanup
import src.cleanupoutliers as cleanoutliers
properties = pd.read_csv('..\data\cleaned_output_2.csv', skip_blank_lines=True)

# Define houses df
houses = properties[properties['type of property'] == 'HOUSE']

# Houses Linear

# Remove non-numerical data from houses
houses = cleanup.drop_column(houses, ['type of property', 'subtype of property', 'province', 'locality', 'postalCode'])
print("There are", houses.shape[0], "rows.")
houses = cleanoutliers.clean_outliers(houses)
print("There are", houses.shape[0], "rows.")
display(houses)


# Convert houses to a numpy dataframe
X = cleanup.drop_column(houses, ['price']).to_numpy()
y = houses['price'].to_numpy()


# Import train_test_split
# Apply train_test_split to dataframe
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=133)

# Fit and score model

from sklearn.linear_model import LinearRegression
regressor = LinearRegression()
regressor.fit(X_train, y_train)
print("The score of the training set is", regressor.score(X_train, y_train))
print("The score of the testing set is", regressor.score(X_test, y_test))



# Houses decision tree

# Import modules
from sklearn.tree import DecisionTreeRegressor
from sklearn.metrics import mean_squared_error



# Define houses df & subsets
houses = properties[properties['type of property'] == 'HOUSE']
print("There are", houses.shape[0], "rows.")
houses = cleanoutliers.clean_outliers(houses)
houses = cleanup.drop_column(houses, ['type of property', 'subtype of property', 'province', 'locality', 'postalCode'])
houses = cleanup.drop_column(houses, ['garden area', 'terrace area','number of facades','open fire'])
houses = houses[houses['living area'] != 0]

print("There are", houses.shape[0], "rows.\n")
X = cleanup.drop_column(houses, 'price').to_numpy()
y = houses['price'].to_numpy()

#Split data
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25)


# With GridsearchCV
from sklearn.model_selection import GridSearchCV
regressor = DecisionTreeRegressor()
params = {'max_depth': [ 4, 6, 8, 10, 12], 'min_weight_fraction_leaf': [0.002, 0.003, 0.0035, 0.004, 0.0045, 0.006]}
grid_search = GridSearchCV(regressor, params, cv=10)
grid_search.fit(X_train, y_train)
best_params = grid_search.best_params_
regressor = DecisionTreeRegressor(max_depth=best_params['max_depth'], min_weight_fraction_leaf=best_params['min_weight_fraction_leaf'])
print(best_params)
regressor.fit(X_train, y_train)
print("The score of the training set is", regressor.score(X_train, y_train))
print("The score of the testing set is", regressor.score(X_test, y_test))
houses.head()



import xgboost
print(xgboost.__version__)

# Define houses df & subsets
houses = properties[properties['type of property'] == 'HOUSE']
print("There are", houses.shape[0], "rows.")
houses = cleanoutliers.clean_outliers(houses)
houses = cleanup.drop_column(houses, ['type of property', 'subtype of property', 'province', 'locality', 'postalCode'])
houses = cleanup.drop_column(houses, ['garden area', 'terrace area','living area'])
#houses = houses[houses['living area'] != 0]
print("There are", houses.shape[0], "rows.\n")
X = cleanup.drop_column(houses, 'price').to_numpy()
y = houses['price'].to_numpy()

regressor = xgboost.XGBRegressor()
#regressor.fit(X_train, y_train)
from sklearn.model_selection import RepeatedKFold
from sklearn.model_selection import cross_val_score
cv = RepeatedKFold(n_splits = 10, n_repeats = 3)
scores = cross_val_score(regressor, X_train, y_train, cv=10, n_jobs = 1)


scores = np.absolute(scores)
print('Mean MAE: %.3f (%.3f)' % (scores.mean(), scores.std()))
